Lift check that path exists to trait
1. Is this worth it? My intent is to write tests that call `run`, so I
don't want to be calling out to the filesystem from within that
function. One could argue that this validation is simple enough that it
could be called directly in main, but a) I like keeping main as minimal
as possible, and b) the Sandu interface may change such that I want to
perform more validation of arguments later.

2. Given this direction, it seems like it may be worth defining some
kind of Runtime struct to capture all these "clients" / dependencies.
However, my initial attempt at collecting Terraform and Filesystem into
a struct produced compiler errors about lifetimes, and I don't want to
go into too deep a rabbit hole with that at this time, so I'm deferring
for later.