Read stdin keys synchronously
The previous implementation chewed up CPU to 100% because it would not
block, so the loop would spin and spin.

There is no need to "do anything" (re-draw, or make background IO
requests or whatever) until a key is pressed, so we can draw the UI and
block, waiting for the user to type something. Fixes the CPU spike
problem!

The `keys()` function is really nice, and I wondered how I might
maintain access to that function without using the termion AsyncReader.
They do not ship a SyncReader... because we can just use std::io::Stdin!
We take advantage of a "blanket implementation". The trait that defines
keys(), TermRead, has an implementation for all T: Read, which means
that by simply pulling in that trait, we can call `keys()` on *any* type
that implements Read, which std::io::Stdin does!

:exploding_head: :star_struck: