Abandon Elm-style Msg/Message enum
With the model state now an enum, there isn't any benefit to matching
the enum case, creating a message, then "forgetting" that match and
essentially redoing it to handle a Message/Msg.

Instead, the state handle function (handling a keypress event) returns
an Optional New State.

Once "draw" is updated to draw the new model instead of the old one,
it'll be able to take an immutable borrow of Mdl, which is great. We'll
also keep the TfPlan around (we'll need it probably in the very next
commit).