Highlight the borders on the active pane
I don't particularly like the wide surface area of the NavigationList
API here, but the only alternative I can think of is to have a function
like `fn matches(&self, other: Self) -> bool` which would require
constructing nonsensical instances of the enum in the draw function and
passing them to this matches function. I'd rather encapsulate those in
the enum impl block.